Improve compare chart

+++ db/migrate/20170623094647_create_chart_data5ms.rb
@@ -9,7 +9,7 @@ class CreateChartData5ms < ActiveRecord::Migration[5.0]
       t.decimal :open, precision: 16, scale: 8
       t.decimal :close, precision: 16, scale: 8
       t.decimal :volume, precision: 16, scale: 8
-      t.decimal :quote_volume, precision: 16, scale: 8
+      t.decimal :quote_volume, precision: 17, scale: 8
       t.decimal :weighted_average, precision: 16, scale: 8
 
       t.index [:currency_pair_id, :time_at], name: 'idx_pair_time_at', unique: true
